requirements.txt
streamlit
google-generativeai
python-dotenv
sentence-transformers
groq
supabase
watchdog
pypdf
python-docx
easyocr
fastapi
uvicorn
ollama
flask
torch
transformers
numpy
scikit-learn
scipy
Pillow
opencv-python
pytesseract
